point handling and file location - moved models and admin to match app - added automatic calculation of points
